generate-blueprint: move preparing  to bootstrap
